Feature "맵툴 타일파트 v1.0"

+++ Main/Header/Struct.h
@@ -71,7 +71,7 @@ typedef struct tagTile
 	BYTE		byOption;
 	float		fDamage;
 
-	BYTE		byDrawID;
+	UINT		byDrawID;
 
 }TILE;
